+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-### Changelog
 
+### Change log
+
+* [#153](https://github.com/dblock/slack-gamebot/pull/153): Added `taunt` command to enable users to taunt other users.
 * [#138](https://github.com/dblock/slack-gamebot/issues/138): Fix: reset season with matches lost and no challenges - [@dblock](https://github.com/dblock).
 * [#113](https://github.com/dblock/slack-gamebot/issues/113): Store elo history - [@dblock](https://github.com/dblock).
 * [#129](https://github.com/dblock/slack-gamebot/issues/129): Added longest winning and losing streaks in leaderboard - [@dblock](https://github.com/dblock).